backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
chromadb==0.5.0
openai==1.40.0
anthropic==0.34.0
langchain==0.2.12
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
pypdf==4.3.0
python-dotenv==1.0.1
pyyaml==6.0.1
httpx==0.27.0
pydantic==2.8.0
pydantic-settings==2.4.0
